add StocksEmbeds.py
